ENG-1448: Split region operations into cohesive private modules
Move construction, boolean composition, contour conversion, disk offsets, simplification, and two-sided gap/width machinery without changing function bodies. Keep existing API paths through re-exports and keep the accuracy contract and prepared query module unchanged.

Move private-helper tests beside their implementations; retain cross-operation tests under region::tests. Replay this move separately after ENG-1444 and ENG-1443 behavior changes, preserving their updated bodies in the corresponding destination modules.

Amp-Thread-ID: https://ampcode.com/threads/T-01a07c7f-fc0e-7071-b923-50402d84bf0c
